Правка по комментариям
